refactor: 중복 검사 함수 예외처리 리팩토링
