+++ src/main/java/baseball/BaseballException.java
@@ -9,16 +9,10 @@ public class BaseballException {
     private BaseballService baseballService = new BaseballService();
 
     // 중복 검사 함수
-    public boolean isDuplicationExist(String randomNum) {
-        ArrayList<Character> randomNumList = new ArrayList<Character>();
-        for (int i=0; i<3; i++){
-            randomNumList.add(randomNum.charAt(i));
+    public void isDuplicationExist(String randomNum) {
+        if (baseballService.isDuplicationExist(randomNum)) {
+            throw new IllegalArgumentException();
         }
-        Set<Character> randNumSet = new HashSet<>(randomNumList);
-        if (randNumSet.size()!=3) {
-            return true;
-        }
-        return false;
     }
 
     // 문자열이 모두 숫자로만 이루어져있는지 검사하는 함수
