+++ src/main/java/baseball/BaseballService.java
@@ -2,6 +2,10 @@ package baseball;
 
 import camp.nextstep.edu.missionutils.Randoms;
 
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.Set;
+
 public class BaseballService {
 
     public String getComputerNum() {
@@ -11,6 +15,19 @@ public class BaseballService {
             computer+= Integer.toString(Randoms.pickNumberInRange(1, 9));
         }
         return computer;
+    }
+
+    // 중복 검사 함수
+    public boolean isDuplicationExist(String randomNum) {
+        ArrayList<Character> randomNumList = new ArrayList<Character>();
+        for (int i=0; i<3; i++){
+            randomNumList.add(randomNum.charAt(i));
+        }
+        Set<Character> randNumSet = new HashSet<>(randomNumList);
+        if (randNumSet.size()!=3) {
+            return true;
+        }
+        return false;
     }
 
     // 문자열이 모두 숫자로만 이루어져있는지 검사하는 함수
